KAN-708 refactor: 일반 식품 삭제 api 연동 수정

+++ src/pages/customer/ProductGeneralDetail.jsx
@@ -4,6 +4,7 @@ import { LeftOutlined } from '@ant-design/icons';
 import { useParams } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
 import { Flex, Form, Input, InputNumber, Switch, Card, Row, Col, Typography, Button, Space, message } from 'antd';
+import Swal from 'sweetalert2';
 
 const { Title } = Typography;
 
@@ -77,15 +78,38 @@ const ProductGeneralDetail = () => {
     }
   };
 
-  const onHandleDelete = async () => {
-    try {
-      await deleteProduct(productId);
-      message.success('상품이 성공적으로 삭제되었습니다.');
-      navigate(-1);  // 상품 목록 페이지로 이동
-    } catch (error) {
-      console.error('Error deleting product: ', error);
-      message.error('상품 삭제에 실패했습니다.');
-    }
+  const onHandleDelete = () => {
+    Swal.fire({
+      title: '정말 삭제하시겠습니까?',
+      text: "이 작업은 되돌릴 수 없습니다!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: '삭제하기',
+      cancelButtonText: '취소'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        console.log('삭제할 상품아이디: ', productId)
+        deleteProduct(productId)
+          .then(() => {
+            Swal.fire(
+              '삭제완료!',
+              '상품이 성공적으로 삭제되었습니다.',
+              'success'
+            );
+            navigate(-1);  // 상품 목록 페이지로 이동
+          })
+          .catch((error) => {
+            console.error('Error deleting product: ', error);
+            Swal.fire(
+              '삭제 실패',
+              '상품 삭제에 실패했습니다.',
+              'error'
+            );
+          });
+      }
+    });
   };
 
   const onHandleBackClick = () => {
@@ -97,13 +121,6 @@ const ProductGeneralDetail = () => {
       <Flex gap="small" justify="flex-start" align="center" style={{ width: 'fit-content' }}>
         <LeftOutlined onClick={onHandleBackClick}/>
         <Title level={3}>식품 등록</Title>
-        {/* <Breadcrumb style={{ marginBottom: '20px' }}
-          items={[
-            { title: 'Main' },
-            { title: selectedFoodType === '일반식품' ? '일반식품관리' : '친환경식품관리' },
-            { title: `${selectedFoodType} 등록` },
-          ]}
-        /> */}
       </Flex>
       <Form form={form} layout="vertical">
         <Row gutter={16}>
@@ -143,9 +160,6 @@ const ProductGeneralDetail = () => {
                   <Form.Item name="baseDiscountRate" label="기본 할인율 (%)" style={formItemStyle}>
                     <InputNumber style={{ ...inputStyle, width: '100%' }} />
                   </Form.Item>
-                  {/* <Form.Item name="regularDiscountRate" label="정기 배송 할인율 (%)" style={formItemStyle}>
-                    <InputNumber style={{ ...inputStyle, width: '100%' }} />
-                  </Form.Item> */}
                   <Form.Item
                     noStyle
                     shouldUpdate={(prevValues, currentValues) => prevValues.isRegularSale !== currentValues.isRegularSale}
